the spotify token is now persistant on refresh
